Re-implement EnabledCallback feature in property presenter
This feature was lost in the migration to the new property presenter. It
allows a property to specify the name of another property or method of
the class which returns a bool. If the callback returns false, the
property input component in the GUI will be disabled/insensitive.

Personally, I think that this is something of an anti-feature and we
will come to regret adding this back in. It's already possible to
emulate this behaviour via DisplayType.SubModel.