Prevent early disposal of widgets in properties GUI
The properties GUI is completely redrawn/recreated whenever one of the
properties is changed. One cause of a "change" being detected is when
focus leaves an entry widget. This could happen, for example, when the
user is typing into an entry and clicks on a dropdown. If this happens,
we cannot dispose of the entire property grid (and therefore the
clicked-on dropdown), because the widgets in the grid may have pending
events. Early disposal in this situation can cause crashes.

My solution here, while not ideal, is to delay disposing of these
widgets until we dispose of the entire property view (ie when the user
clicks away to another node). At this point in time it should be safe to
dispose of them. The problem is that when the user clicks out of an
entry onto a dropdown, the old dropdown still receives the event and
pops up, and selecting something in the old popup has no effect (because
we've already detached the event handlers for the old dropdown).