Fixed drag-and-drop in GUI on MacOS
Previously, dragging a model to a different parent was not working,
because the suggested action returned from the gdk-quartz backend (move
| copy | link) was different to the suggested action from the x11
backend (copy). There was a bug in the code which checks the action - it
was performing a straight equality check, and didn't account for the
possibility of multiple bit fields being set in this value.

I've corrected this particular problem, but the behaviour will still be
slightly different between quartz (MacOS) and x11 (Linux), and
presumably, windows as well; holding shift while dragging in x11 will
result in a "move" action, whereas dragging without any keypress will
result in a copy. Under quartz, dragging without any keypress will
result in a move. I'm going to accept these differences as acceptable -
hopefully the different behaviours make sense to the users of the
respective platforms.